Fix #4
Since the language setting of the FritzBox determines the csv output, we cannot simply assume that the output is in English.
Thanks @fur6y for catching this! - https://github.com/lesander/fritzbox.js/issues/4

The second line of the CSV file contains the column titles.
DE: Typ  ; Datum ; Name ; Rufnummer        ; Nebenstelle ; Eigene Rufnummer ; Dauer
EN: Type ; Date  ; Name ; Telephone number ; Extension   ; Telephone Number ; Duration

We could count the columns and name them ourselves, or replace the second line of the CSV
file and replace it with the English format. I've implemented the latter for now,
this is however not the cleanest way of achieving multi-language support.